Refactor objects structure.
