+++ src/main/scala/com/evo/bootcamp/quiz/dao/models/Answer.scala
@@ -1,0 +1,5 @@
+package com.evo.bootcamp.quiz.dao.models
+
+case class Answer(id: Int, text: String, isRight: Boolean) {
+  override def toString: String = s"$text"
+}
