+++ src/main/scala/com/evo/bootcamp/quiz/dao/models/QuestionWithAnswer.scala
@@ -1,3 +1,3 @@
 package com.evo.bootcamp.quiz.dao.models
 
-final case class QuestionWithAnswer(id: Int, text: String, answerId: Int, answerText: String, answerIsRight: Boolean)
+final case class QuestionWithAnswer(id: Int, text: String, answer: Answer)
