--- src/main/scala/com/evo/bootcamp/quiz/dto/Answer.scala
@@ -1,8 +1,0 @@
-package com.evo.bootcamp.quiz.dto
-
-case class Answer(id: Int, text: String, isRight: Boolean) extends Comparable[Answer] {
-  override def toString: String = s"$text"
-
-  override def compareTo(answer: Answer): Int =
-    this.id - answer.id
-}
