+++ src/main/scala/com/evo/bootcamp/quiz/dto/AnswerDto.scala
@@ -1,0 +1,8 @@
+package com.evo.bootcamp.quiz.dto
+
+case class AnswerDto(id: Int, text: String, isRight: Boolean) extends Comparable[AnswerDto] {
+  override def toString: String = s"$text"
+
+  override def compareTo(answer: AnswerDto): Int =
+    this.id - answer.id
+}
