--- src/main/scala/com/evo/bootcamp/quiz/dto/BotMessage.scala
@@ -1,3 +1,0 @@
-package com.evo.bootcamp.quiz.dto
-
-final case class BotMessage(chat: Chat, text: Option[String])
