--- src/main/scala/com/evo/bootcamp/quiz/dto/BotUpdate.scala
@@ -1,3 +1,0 @@
-package com.evo.bootcamp.quiz.dto
-
-final case class BotUpdate(update_id: Long, message: Option[BotMessage], callback_query: Option[CallbackQuery])
