--- src/main/scala/com/evo/bootcamp/quiz/dto/CallbackQuery.scala
@@ -1,3 +1,0 @@
-package com.evo.bootcamp.quiz.dto
-
-final case class CallbackQuery(from: User, data: Option[String])
