--- src/main/scala/com/evo/bootcamp/quiz/dto/GameResult.scala
@@ -1,3 +1,0 @@
-package com.evo.bootcamp.quiz.dto
-
-final case class GameResult(rightAnswersAmount: Int, totalAmount: Int)
