+++ src/main/scala/com/evo/bootcamp/quiz/dto/GameResultDto.scala
@@ -1,0 +1,3 @@
+package com.evo.bootcamp.quiz.dto
+
+final case class GameResultDto(rightAnswersAmount: Int, totalAmount: Int)
