--- src/main/scala/com/evo/bootcamp/quiz/dto/MessageResponse.scala
@@ -1,5 +1,0 @@
-package com.evo.bootcamp.quiz.dto
-
-case class MessageResponse(ok: Boolean, result: Result)
-
-case class Result(message_id: Long)
