+++ src/main/scala/com/evo/bootcamp/quiz/dto/QuestionDto.scala
@@ -1,0 +1,5 @@
+package com.evo.bootcamp.quiz.dto
+
+case class QuestionDto(id: Int, text: String, answers: List[AnswerDto], rightAnswer: Option[AnswerDto], var userAnswer: Option[AnswerDto]) {
+  override def toString: String = s"$text"
+}
