--- src/main/scala/com/evo/bootcamp/quiz/dto/User.scala
@@ -1,4 +1,0 @@
-package com.evo.bootcamp.quiz.dto
-
-final case class User(id: Long)
-
