+++ src/main/scala/com/evo/bootcamp/quiz/dto/api/BotResponse.scala
@@ -1,3 +1,3 @@
-package com.evo.bootcamp.quiz.dto
+package com.evo.bootcamp.quiz.dto.api
 
 final case class BotResponse[T](ok: Boolean, result: T)
