+++ src/main/scala/com/evo/bootcamp/quiz/dto/api/BotUpdateMessage.scala
@@ -1,0 +1,3 @@
+package com.evo.bootcamp.quiz.dto.api
+
+final case class BotUpdateMessage(update_id: Long, message: Option[Message], callback_query: Option[CallbackQuery])
