+++ src/main/scala/com/evo/bootcamp/quiz/dto/api/Chat.scala
@@ -1,0 +1,3 @@
+package com.evo.bootcamp.quiz.dto.api
+
+final case class Chat(id: Long)
