+++ src/main/scala/com/evo/bootcamp/quiz/dto/api/InlineKeyboardButton.scala
@@ -1,3 +1,3 @@
-package com.evo.bootcamp.quiz.dto
+package com.evo.bootcamp.quiz.dto.api
 
 final case class InlineKeyboardButton(text: String, callback_data: String)
