+++ src/main/scala/com/evo/bootcamp/quiz/dto/api/Message.scala
@@ -1,0 +1,3 @@
+package com.evo.bootcamp.quiz.dto.api
+
+final case class Message(chat: Chat, text: Option[String])
