+++ src/main/scala/com/evo/bootcamp/quiz/mapper/Mappings.scala
@@ -1,0 +1,8 @@
+package com.evo.bootcamp.quiz.mapper
+
+import com.evo.bootcamp.quiz.dao.models.Answer
+import com.evo.bootcamp.quiz.dto.AnswerDto
+
+object Mappings {
+  implicit def toAnswerDto(answer: Answer) = AnswerDto(answer.id, answer.text, answer.isRight)
+}
